Fix /analysis for Pokemon targets
Fixed the "previews" target link. Originally set as https://www.pokebattle.com/pokedex/ when the real url is https://www.pokebattle.com/dex/
Also removed the generation parameter from the Pokemon search in an attempt to get the command to actually work, since it does not work on the test server.